+++ src/include/cdb/cdbcopy.h
@@ -70,6 +70,6 @@ void		cdbCopySendDataToAll(CdbCopy *c, const char *buffer, int nbytes);
 void		cdbCopySendData(CdbCopy *c, int target_seg, const char *buffer, int nbytes);
 bool		cdbCopyGetData(CdbCopy *c, bool cancel, uint64 *rows_processed);
 int			cdbCopyEnd(CdbCopy *c);
-int			cdbCopyEndAndFetchRejectNum(CdbCopy *c, int *total_rows_completed);
+int			cdbCopyEndAndFetchRejectNum(CdbCopy *c, int64 *total_rows_completed);
 
 #endif   /* CDBCOPY_H */
